move lib/client to _utilities/apollo_client
